Rename path_manager to fso_expansion
