+++ pyproject.toml
@@ -9,8 +9,8 @@ readme = "README.md"
 python = "^3.12"
 
 [tool.poetry.scripts]
-pcmkdir = 'path_chronicle.path_manager:create_dir_entry'
-pctouch = 'path_chronicle.path_manager:create_file_entry'
+pcmkdir = 'path_chronicle.fso_expansion:create_dir_entry'
+pctouch = 'path_chronicle.fso_expansion:create_file_entry'
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.2"
